pyproject.toml
[project]
name = "finances-api"
version = "0.1.0"
description = ""
authors = [
    {name = "diioliiver",email = "diego.feitosa.oliveira@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "dependency-injector (>=4.47.1,<5.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "fastapi-pagination (>=0.13.2,<0.14.0)",
]

[tool.poetry]
name = "finances-api"
version = "0.1.0"
description = ""
authors = ["diioliiver <diego.feitosa.oliveira@hotmail.com>"]

[[tool.poetry.packages]]
include = "application"
from = "src"

[[tool.poetry.packages]]
include = "domain"
from = "src"

[[tool.poetry.packages]]
include = "infra"
from = "src"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'uvicorn main:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
check_arch = 'lint-imports --config .importlinter  --no-cache'

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR6201', 'E501']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
ruff = "^0.11.9"
import-linter = "^2.3"
pre-commit = "^4.1.0"